Refactor profile.rb
